Moved code into utils and added extra inheritance tokens
